backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
numpy==2.1.3
faster-whisper==1.0.3
webrtcvad==2.0.10
pyannote.audio==3.3.2
soundfile==0.12.1
torch==2.5.1
python-multipart==0.0.12
pypdf==5.1.0
python-docx==1.1.2
transformers==4.46.2
